pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "googletrans==4.0.0rc1",
    "gtts>=2.5.4",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "requests>=2.32.4",
    "streamlit>=1.47.0",
]
